Refactor to move Camera and Teleport component to separate MonoBehaviour
